fixed Dockerfile for mcr2007

+++ stop.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
-jobid=`cat jobid`
-echo "running qdel $jobid"
-qdel $jobid
+if [ -f jobid ]; then
+	jobid=`cat jobid`
+	echo "running qdel $jobid"
+	qdel $jobid
+fi
+
+if [ -f pid ]; then
+	kill $(cat pid)
+fi
